Plot Summary Statistics of Stochastic Character Map Transitions

Creates a two-panel plot to visualize the results from a list of stochastic character mapping simulations. The function generates a consolidated boxplot for transition counts and a staggered density plot (ridgeline plot) for transition ages.

plot_transition_summary(simmap_summaries, cols = NULL)

Arguments

simmap_summaries

A list of data frames, where each data frame is the result of applying summarize_single_simmap to a "simmap" object.

cols

An optional vector of character strings specifying colors for the plots. If NULL, a default color palette is used. Colors are recycled if there are more transition types than colors provided.

Details

This function produces a visualization of simulation results from stochastic character mapping. It is designed to be used after processing a list of "simmap" objects with lapply(simmaps, summarize_single_simmap).

The graphical output consists of two main panels:

Left Panel: A single plot containing vertical boxplots for the count of each transition type, summarized across all simulations. Semi-transparent points are overlaid with a small amount of jitter to show the distribution of individual data points.

Right Panel: A ridgeline plot showing the distribution of transition ages.

Examples

if (FALSE) { # \dontrun{
data(primates)
phy <- primates[[1]]
phy <- multi2di(phy)
data <- primates[[2]]

##run corhmm
MK <- corHMM(phy, data, 1)

##get simmap from corhmm solution
model <- MK$solution
simmaps <- makeSimmap(tree=phy, data=data, model=model, rate.cat=1, nSim=100, nCores=1)

# get the summary
simmap_summaries <- lapply(simmaps, summarize_single_simmap)
summary_df <- summarize_transition_stats(simmap_summaries)
print(summary_df)

plot_transition_summary(simmap_summaries)
} # }