Perform K-Fold Cross-Validation for corHMM Models

This function performs k-fold cross-validation on a given corHMM model by dividing the data into k equally sized subsets. The function evaluates model performance across multiple lambda regularization values, if provided. Optionally, it can save the trained models for each fold and return the cross-validation results.

kFoldCrossValidation(corhmm_obj, k, lambdas = NULL, return_model = TRUE, 
save_model_dir = NULL, model_name = NULL)

Arguments

corhmm_obj

A corHMM object that contains a fitted model.

k

An integer specifying the number of folds to divide the data into for cross-validation.

lambdas

A numeric vector of lambda regularization values to evaluate during cross-validation. If NULL, the lambda value from corhmm_obj will be used. Defaults to NULL.

return_model

A logical value indicating whether to return the trained models for each fold. Defaults to TRUE.

save_model_dir

A character string specifying the directory to save the trained models for each fold. If NULL, models will not be saved. Defaults to NULL.

model_name

A character string specifying the base name for saved model files. If NULL, a default name "corhmm.obj" is used. Defaults to NULL.

Details

The function splits the data into k folds and trains a separate corHMM model for each fold by leaving one fold out as the test set. The remaining folds are used for training the model. The performance of the model is evaluated on the test set using a divergence-based (Jensen-Shannon Divergence) scoring method. Evaluations are based on estimating the tips which were removed for that particular fold given the newly fitted model.

The function supports evaluating models across different lambda regularization values. If lambdas are provided, models are trained and evaluated for each lambda value. The results, including the models (if return_model = TRUE) and cross-validation scores, are returned as a list.

Value

A list of cross-validation results, including the following components:

models

A list of the trained models for each fold (if return_model = TRUE).

scores

A numeric vector of the cross-validation scores for each fold.

averageScore

The average cross-validation score across all folds.

Author

James D. Boyko

Examples

# \donttest{
# run your dredge
data(primates)
phy <- multi2di(primates[[1]])
data <- primates[[2]]
dredge_fits <- corHMMDredge(phy = phy, data = data, 
  max.rate.cat = 1, pen.type = "l1", 
  root.p = "maddfitz", lambda = 1, nstarts = 10, n.cores = 10)
#> Beginning SA dredge...
#> SA parameters: max_iter = 1000 , init_temp = 2 , cooling = 0.95 
#> 
#> === RATE CATEGORY 1 ===
#> Warning: Branch lengths of 0 detected. Adding 1.49011611938477e-08
#> Initial AIC : 101.5139 
#>     0|0 1|0 0|1 1|1
#> 0|0  NA   3   5  NA
#> 1|0   1  NA  NA   7
#> 0|1   2  NA  NA   8
#> 1|1  NA   4   6  NA
#> 
#> Starting SA with AIC = 101.514 
#> Restart strategy: fixed_steps 
#> Iter 1 - New AIC: 97.647 - Best AIC: 97.647 
#>  Move: drop - Temp: 1.9 - Steps since restart: 1 
#>  Index Matrix:
#>     0|0 1|0 0|1 1|1
#> 0|0  NA   2   4  NA
#> 1|0  NA  NA  NA   6
#> 0|1   1  NA  NA  NA
#> 1|1  NA   3   5  NA
#> 
#> Iter 2 - New AIC: 95.647 - Best AIC: 95.647 
#>  Move: merge - Temp: 1.805 - Steps since restart: 2 
#>  Index Matrix:
#>     0|0 1|0 0|1 1|1
#> 0|0  NA   2   3  NA
#> 1|0  NA  NA  NA   5
#> 0|1   1  NA  NA  NA
#> 1|1  NA   5   4  NA
#> 
#> Iter 3 - New AIC: 93.647 - Best AIC: 93.647 
#>  Move: merge - Temp: 1.7147 - Steps since restart: 3 
#>  Index Matrix:
#>     0|0 1|0 0|1 1|1
#> 0|0  NA   4   2  NA
#> 1|0  NA  NA  NA   4
#> 0|1   1  NA  NA  NA
#> 1|1  NA   4   3  NA
#> 
#> Iter 4 - New AIC: 91.647 - Best AIC: 91.647 
#>  Move: drop - Temp: 1.629 - Steps since restart: 4 
#>  Index Matrix:
#>     0|0 1|0 0|1 1|1
#> 0|0  NA  NA   2  NA
#> 1|0  NA  NA  NA  NA
#> 0|1   1  NA  NA  NA
#> 1|1  NA  NA   3  NA
#> 
#> Iter 5 - New AIC: 91.141 - Best AIC: 91.141 
#>  Move: drop - Temp: 1.5476 - Steps since restart: 5 
#>  Index Matrix:
#>     0|0 1|0 0|1 1|1
#> 0|0  NA  NA  NA  NA
#> 1|0  NA  NA  NA  NA
#> 0|1   1  NA  NA  NA
#> 1|1  NA  NA   2  NA
#> 
#> Iter 7 - New AIC: 95.141 - Best AIC: 91.141 
#>  Move: free - Temp: 1.3967 - Steps since restart: 7 
#>  Index Matrix:
#>     0|0 1|0 0|1 1|1
#> 0|0  NA  NA  NA  NA
#> 1|0   4  NA  NA  NA
#> 0|1   1  NA  NA  NA
#> 1|1  NA   3   2  NA
#> 
#> Iter 9 - New AIC: 93.141 - Best AIC: 91.141 
#>  Move: merge - Temp: 1.2605 - Steps since restart: 9 
#>  Index Matrix:
#>     0|0 1|0 0|1 1|1
#> 0|0  NA  NA  NA  NA
#> 1|0   3  NA  NA  NA
#> 0|1   1  NA  NA  NA
#> 1|1  NA   3   2  NA
#> 
#> Iter 12 - New AIC: 93.141 - Best AIC: 91.141 
#>  Move: drop - Temp: 1.0807 - Steps since restart: 12 
#>  Index Matrix:
#>     0|0 1|0 0|1 1|1
#> 0|0  NA  NA  NA  NA
#> 1|0   3  NA  NA  NA
#> 0|1   1  NA  NA  NA
#> 1|1  NA  NA   2  NA
#> 
#> Iter 13 - New AIC: 91.773 - Best AIC: 91.141 
#>  Move: merge - Temp: 1.0267 - Steps since restart: 13 
#>  Index Matrix:
#>     0|0 1|0 0|1 1|1
#> 0|0  NA  NA  NA  NA
#> 1|0   2  NA  NA  NA
#> 0|1   1  NA  NA  NA
#> 1|1  NA  NA   2  NA
#> 
#> RESTART 1 at iteration 33 ( fixed_steps ) - Back to a previous model AIC : 93.647 
#> Iter 34 - New AIC: 93.141 - Best AIC: 91.141 
#>  Move: merge - Temp: 0.3496 - Steps since restart: 1 
#>  Index Matrix:
#>     0|0 1|0 0|1 1|1
#> 0|0  NA   3   3  NA
#> 1|0  NA  NA  NA   3
#> 0|1   1  NA  NA  NA
#> 1|1  NA   3   2  NA
#> 
#> Iter 37 - New AIC: 91.792 - Best AIC: 91.141 
#>  Move: drop - Temp: 0.2998 - Steps since restart: 4 
#>  Index Matrix:
#>     0|0 1|0 0|1 1|1
#> 0|0  NA  NA   3  NA
#> 1|0  NA  NA  NA   3
#> 0|1   1  NA  NA  NA
#> 1|1  NA  NA   2  NA
#> 
#> Iter 40 - New AIC: 91.647 - Best AIC: 91.141 
#>  Move: drop - Temp: 0.257 - Steps since restart: 7 
#>  Index Matrix:
#>     0|0 1|0 0|1 1|1
#> 0|0  NA  NA   3  NA
#> 1|0  NA  NA  NA  NA
#> 0|1   1  NA  NA  NA
#> 1|1  NA  NA   2  NA
#> 
#> Iter 45 - New AIC: 91.544 - Best AIC: 91.141 
#>  Move: merge - Temp: 0.1989 - Steps since restart: 12 
#>  Index Matrix:
#>     0|0 1|0 0|1 1|1
#> 0|0  NA  NA   2  NA
#> 1|0  NA  NA  NA  NA
#> 0|1   1  NA  NA  NA
#> 1|1  NA  NA   2  NA
#> 
#> RESTART 2 at iteration 65 ( fixed_steps ) - Back to a previous model AIC : 93.647 
#> Iter 67 - New AIC: 93.141 - Best AIC: 91.141 
#>  Move: drop - Temp: 0.0643 - Steps since restart: 2 
#>  Index Matrix:
#>     0|0 1|0 0|1 1|1
#> 0|0  NA   3  NA  NA
#> 1|0  NA  NA  NA   3
#> 0|1   1  NA  NA  NA
#> 1|1  NA   3   2  NA
#> 
#> Iter 70 - New AIC: 93.141 - Best AIC: 91.141 
#>  Move: drop - Temp: 0.0552 - Steps since restart: 5 
#>  Index Matrix:
#>     0|0 1|0 0|1 1|1
#> 0|0  NA  NA  NA  NA
#> 1|0  NA  NA  NA   3
#> 0|1   1  NA  NA  NA
#> 1|1  NA  NA   2  NA
#> 
#> Iter 75 - New AIC: 91.977 - Best AIC: 91.141 
#>  Move: merge - Temp: 0.0427 - Steps since restart: 10 
#>  Index Matrix:
#>     0|0 1|0 0|1 1|1
#> 0|0  NA  NA  NA  NA
#> 1|0  NA  NA  NA   2
#> 0|1   1  NA  NA  NA
#> 1|1  NA  NA   2  NA
#> 
#> RESTART 3 at iteration 95 ( fixed_steps ) - Back to a previous model AIC : 91.544 
#> RESTART 4 at iteration 115 ( fixed_steps ) - Back to a previous model AIC : 95.141 
#> Iter 121 - New AIC: 93.141 - Best AIC: 91.141 
#>  Move: drop - Temp: 0.004 - Steps since restart: 6 
#>  Index Matrix:
#>     0|0 1|0 0|1 1|1
#> 0|0  NA  NA  NA  NA
#> 1|0  NA  NA  NA  NA
#> 0|1   1  NA  NA  NA
#> 1|1  NA   3   2  NA
#> 
#> RESTART 5 at iteration 141 ( fixed_steps ) - Back to a previous model AIC : 91.544 
#> Best AIC for Rate Class 1 after SA: 91.141 
#> SA iterations: 149 
#> SA acceptance rate: 0.274 
#> 
#> Done.

# select the best model and run CV
model_table <- getModelTable(dredge_fits)
dredge_model <- dredge_fits[[which.min(model_table$dAIC)]]
k_fold_res <- kFoldCrossValidation(dredge_model,
k = 5, lambdas = c(0,0.25,0.5,0.75,1))
#> Evaluating lambda = 0 
#> Warning: Branch lengths of 0 detected. Adding 1.49011611938477e-08
#> Warning: NA/NaN function evaluation
#> Fold 0 Score: 0.04513702 
#> Warning: Branch lengths of 0 detected. Adding 1.49011611938477e-08
#> Warning: NA/NaN function evaluation
#> Fold 1 Score: 0.03252338 
#> Warning: Branch lengths of 0 detected. Adding 1.49011611938477e-08
#> Warning: NA/NaN function evaluation
#> Fold 2 Score: 0.02020864 
#> Warning: Branch lengths of 0 detected. Adding 1.49011611938477e-08
#> Warning: NA/NaN function evaluation
#> Fold 3 Score: 0.02721126 
#> Warning: Branch lengths of 0 detected. Adding 1.49011611938477e-08
#> Warning: NA/NaN function evaluation
#> Fold 4 Score: 0.02740597 
#> Average Cross-Validation Score: 0.03049725 
#> Evaluating lambda = 0.25 
#> Warning: Branch lengths of 0 detected. Adding 1.49011611938477e-08
#> Warning: NA/NaN function evaluation
#> Fold 0 Score: 0.04504975 
#> Warning: Branch lengths of 0 detected. Adding 1.49011611938477e-08
#> Warning: NA/NaN function evaluation
#> Fold 1 Score: 0.03252082 
#> Warning: Branch lengths of 0 detected. Adding 1.49011611938477e-08
#> Warning: NA/NaN function evaluation
#> Fold 2 Score: 0.02018284 
#> Warning: Branch lengths of 0 detected. Adding 1.49011611938477e-08
#> Warning: NA/NaN function evaluation
#> Fold 3 Score: 0.02718753 
#> Warning: Branch lengths of 0 detected. Adding 1.49011611938477e-08
#> Warning: NA/NaN function evaluation
#> Fold 4 Score: 0.02741115 
#> Average Cross-Validation Score: 0.03047042 
#> Evaluating lambda = 0.5 
#> Warning: Branch lengths of 0 detected. Adding 1.49011611938477e-08
#> Warning: NA/NaN function evaluation
#> Fold 0 Score: 0.04496942 
#> Warning: Branch lengths of 0 detected. Adding 1.49011611938477e-08
#> Warning: NA/NaN function evaluation
#> Fold 1 Score: 0.03251816 
#> Warning: Branch lengths of 0 detected. Adding 1.49011611938477e-08
#> Warning: NA/NaN function evaluation
#> Fold 2 Score: 0.02016062 
#> Warning: Branch lengths of 0 detected. Adding 1.49011611938477e-08
#> Warning: NA/NaN function evaluation
#> Fold 3 Score: 0.02716733 
#> Warning: Branch lengths of 0 detected. Adding 1.49011611938477e-08
#> Warning: NA/NaN function evaluation
#> Fold 4 Score: 0.02742075 
#> Average Cross-Validation Score: 0.03044726 
#> Evaluating lambda = 0.75 
#> Warning: Branch lengths of 0 detected. Adding 1.49011611938477e-08
#> Warning: NA/NaN function evaluation
#> Fold 0 Score: 0.04489525 
#> Warning: Branch lengths of 0 detected. Adding 1.49011611938477e-08
#> Warning: NA/NaN function evaluation
#> Fold 1 Score: 0.03251539 
#> Warning: Branch lengths of 0 detected. Adding 1.49011611938477e-08
#> Warning: NA/NaN function evaluation
#> Fold 2 Score: 0.02014147 
#> Warning: Branch lengths of 0 detected. Adding 1.49011611938477e-08
#> Warning: NA/NaN function evaluation
#> Fold 3 Score: 0.02715007 
#> Warning: Branch lengths of 0 detected. Adding 1.49011611938477e-08
#> Warning: NA/NaN function evaluation
#> Fold 4 Score: 0.02743372 
#> Average Cross-Validation Score: 0.03042718 
#> Evaluating lambda = 1 
#> Warning: Branch lengths of 0 detected. Adding 1.49011611938477e-08
#> Warning: NA/NaN function evaluation
#> Fold 0 Score: 0.04482662 
#> Warning: Branch lengths of 0 detected. Adding 1.49011611938477e-08
#> Warning: NA/NaN function evaluation
#> Fold 1 Score: 0.0325125 
#> Warning: Branch lengths of 0 detected. Adding 1.49011611938477e-08
#> Warning: NA/NaN function evaluation
#> Fold 2 Score: 0.02012498 
#> Warning: Branch lengths of 0 detected. Adding 1.49011611938477e-08
#> Warning: NA/NaN function evaluation
#> Fold 3 Score: 0.02713534 
#> Warning: Branch lengths of 0 detected. Adding 1.49011611938477e-08
#> Warning: NA/NaN function evaluation
#> Fold 4 Score: 0.02744932 
#> Average Cross-Validation Score: 0.03040975 

cv_table <- getCVTable(k_fold_res)

# choose lowest average score for lambda
print(cv_table)
#> $score_table
#>          Lambda:0 Lambda:0.25 Lambda:0.5 Lambda:0.75   Lambda:1
#> Fold:0 0.04513702  0.04504975 0.04496942  0.04489525 0.04482662
#> Fold:1 0.03252338  0.03252082 0.03251816  0.03251539 0.03251250
#> Fold:2 0.02020864  0.02018284 0.02016062  0.02014147 0.02012498
#> Fold:3 0.02721126  0.02718753 0.02716733  0.02715007 0.02713534
#> Fold:4 0.02740597  0.02741115 0.02742075  0.02743372 0.02744932
#> 
#> $avg_scores
#>    Lambda:0 Lambda:0.25  Lambda:0.5 Lambda:0.75    Lambda:1 
#>  0.03049725  0.03047042  0.03044726  0.03042718  0.03040975 
#> 
# }